Redirect Qt logging into Rust's tracing
We use qInstallMessageHandler() to install a custom handler, which
is, in fact, implemented in Rust and converts the Qt messages into
tracing's events.

It's not perfect, the format of the captured Qt log message could be
improved, but it's a good start to get complete view of what's going
on between the Rust backend and QML frontend.